web-game/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-socketio>=5.10.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
aiosqlite>=0.19.0
Pillow>=10.0.0
pydantic>=2.4.0
websockets>=12.0
cors>=1.0.1